services/patient-data/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
boto3==1.34.131
psycopg2-binary==2.9.9
